@@ -1,95 +1,176 @@
 name: bili_you
 description: A new Flutter project.
 
-publish_to: 'none'
+# The following line prevents the package from being accidentally published to
+# pub.dev using `flutter pub publish`. This is preferred for private packages.
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 1.1.515
+version: 1.1.5+15
 
 environment:
-  sdk: '>3.0.0-417.0.dev'
+  sdk: '>=3.0.0-417.0.dev'
 
 dependencies:
   flutter:
     sdk: flutter
 
   # get
-  get: 4.6.5
+  get: ^4.6.5
 
   # Material You 动态配色
-  dynamic_color: 1.5.2
+  dynamic_color: ^1.6.3
 
-  # 网络相关
-  dio: 5.3.2
-  dio_cookie_manager: 3.1.1
-  cookie_jar: 4.0.8
-  connectivity_plus: 5.0.2
+  # 网络相关 
+  dio: ^5.3.2
+  dio_cookie_manager: ^3.1.0
+  cookie_jar: ^4.0.8
+  gt3_flutter_plugin: ^0.0.8
+  flutter_cache_manager: ^3.3.1
+  protobuf: ^3.0.0
+  connectivity_plus: ^4.0.2
 
-  # 本地存储
-  shared_preferences: 2.2.2
-  path_provider: 2.1.1
-  hive: 2.2.3
-  hive_flutter: 1.1.0
+  # 加解密
+  encrypt: ^5.0.1
+  crypto: ^3.0.3
 
-  # 视频播放
-  video_player: 2.8.1
-  media_kit: 1.1.10
-  media_kit_video: 1.2.4
-  media_kit_libs_android_video: 1.1.3
-  media_kit_libs_ios_video: 1.1.3
-  media_kit_libs_linux_video: 1.1.3
-
-  # 状态管理
-  flutter_bloc: 8.1.3
-  equatable: 2.0.5
-
-  # 弹幕
-  flutter_ns_danmaku:
-    git:
+  # 文件路径
+  path: ^1.8.3
+  path_provider: ^2.0.14
+  easy_refresh: ^3.3.2+1
+  
+  # 控件
+  ns_danmaku:
+    git: 
       url: https://github.com/lucinhu/flutter_ns_danmaku.git
+  cached_network_image: ^3.2.3
+  photo_view: ^0.14.0  
+  card_swiper: ^3.0.1  
 
-  # UI 组件
-  flutter_svg: 2.0.9
-  extended_image: 8.2.0
-  flutter_smart_dialog: 4.9.6
-  pull_to_refresh_flutter3: 2.0.2
+  # 播放器
+  media_kit: ^1.1.1
+  media_kit_libs_android_video: ^1.3.0
+  media_kit_libs_ios_video: ^1.0.5
+  media_kit_libs_linux: ^1.0.2
+  media_kit_libs_macos_video: ^1.0.6
+  media_kit_libs_windows_video: ^1.0.5
+  media_kit_native_event_loop: ^1.0.6
+  media_kit_video: ^1.1.1
 
-  # 工具类
-  logger: 2.0.2+1
-  device_info_plus: 9.1.1
-  package_info_plus: 4.2.0
-  url_launcher: 6.2.2
+  # webview
+  webview_cookie_manager: ^2.0.6
+  webview_flutter: ^4.2.2
 
-  # 平台特定
-  window_manager: ^0.3.7
-  screen_retriever: ^0.1.9
+  # 保持常亮
+  wakelock_plus: ^1.1.1
+
+  # 设备&包信息
+  device_info_plus: ^9.0.3
+  package_info_plus: ^4.1.0
+
+  # url跳转
+  url_launcher: ^6.1.12
+  
+  # hive
+  hive: ^2.2.3
+  hive_flutter: ^1.1.0
+
+  # 分享
+  share_plus: ^7.1.0
+
+  # scheme 
+  appscheme: ^1.0.8
+
+  # 音量控制
+  volume_controller: ^2.0.6
+
+  # 亮度控制
+  screen_brightness: ^0.2.2
+
+  # 自动旋转屏幕
+  auto_orientation: ^2.3.1
+
+  # 二维码
+  qr_flutter: ^4.1.0
+
+  #Toast
+  fluttertoast: ^8.2.2
+
+  #圖片保存
+  image_gallery_saver: ^2.0.3
+
+# dependency_overrides:
+  # cookie_jar:
+  #   git:
+  #     url: https://github.com/dkaera/cookie_jar.git
+  #     ref: fix/cast_exception
+
+#  wakelock_windows:
+#    git:
+#      url: https://github.com/timsneath/wakelock
+#      ref: win32-v5
+#      path: wakelock_windows
+
+  # media_kit:
+  #   git: 
+  #     url: https://github.com/alexmercerind/media_kit.git
+  #     path: media_kit
+  # media_kit_libs_android_video:
+  #   git: 
+  #     url: https://github.com/alexmercerind/media_kit.git
+  #     path: media_kit_libs_android_video
+  # media_kit_libs_ios_video: 
+  #   git: 
+  #     url: https://github.com/alexmercerind/media_kit.git
+  #     path: media_kit_libs_ios_video
+  # media_kit_libs_linux:
+  #   git: 
+  #     url: https://github.com/alexmercerind/media_kit.git
+  #     path: media_kit_libs_linux
+  # media_kit_libs_macos_video:
+  #   git: 
+  #     url: https://github.com/alexmercerind/media_kit.git
+  #     path: media_kit_libs_macos_video
+  # media_kit_libs_windows_video:
+  #   git: 
+  #     url: https://github.com/alexmercerind/media_kit.git
+  #     path: media_kit_libs_windows_video
+  # media_kit_video:
+  #   git: 
+  #     url: https://github.com/alexmercerind/media_kit.git
+  #     path: media_kit_video
+  # media_kit_native_event_loop:
+  #   git: 
+  #     url: https://github.com/alexmercerind/media_kit.git
+  #     path: media_kit_native_event_loop
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  flutter_lints: ^3.0.0
-  hive_generator: ^2.0.1
-  build_runner: ^2.4.7
+  build_runner: ^2.4.4
+  test: ^1.24.1
+  flutter_launcher_icons: ^0.13.1
+  flutter_lints: ^2.0.1
 
 flutter:
   uses-material-design: true
   assets:
-    - assets/images/
-    - assets/icon/bili.png
-  fonts:
-    - family: bili
-      fonts:
-        - asset: assets/fonts/bili.ttf
+      - assets/icon/bili.png
 
-flutter_icons:
+flutter_launcher_icons:
   android: true
   ios: true
   image_path: "assets/icon/bili.png"
-  background_color: "#hexcode"
-  theme_color: "#hexcode"
+  adaptive_icon_background: "#ffffff"
+  adaptive_icon_foreground: "assets/icon/bili.png"
+  remove_alpha_ios: true
+  min_sdk_android: 21 # android min sdk min:16, default 21
+  web:
+    generate: true
+    image_path: "assets/icon/bili.png"
+    background_color: "#hexcode"
+    theme_color: "#hexcode"
   windows:
     generate: true
     image_path: "assets/icon/bili.png"
-    icon_size: 48
+    icon_size: 48 # min:48, max:256, default: 48
   macos:
     generate: true
     image_path: "assets/icon/bili.png"
