@@ -68,26 +68,21 @@ class SearchApi {
 
   ///根据keyWord获取搜索建议
   static Future<List<SearchSuggestItem>> getSearchSuggests(
-    {required String keyWord}) async {
-  List<SearchSuggestItem> list = [];
-  var response = await _requestSearchSuggests(keyWord);
-  if (response.code != 0) {
-    throw "getSearchSuggests: code:${response.code}";
-  }
-  if (response.result == null || response.result!.tag == null) {
+      {required String keyWord}) async {
+    List<SearchSuggestItem> list = [];
+    var response = await _requestSearchSuggests(keyWord);
+    if (response.code != 0) {
+      throw "getSearchSuggests: code:${response.code}";
+    }
+    if (response.result == null || response.result!.tag == null) {
+      return list;
+    }
+    for (var i in response.result!.tag!) {
+      list.add(
+          SearchSuggestItem(showWord: i.name ?? "", realWord: i.value ?? ""));
+    }
     return list;
   }
-  for (var i in response.result!.tag!) {
-    // ✅ 修复：使用 keyWordTitleToRawTitle 移除 HTML 标签
-    final cleanName = StringFormatUtils.keyWordTitleToRawTitle(i.name ?? "");
-    
-    list.add(SearchSuggestItem(
-  showWord: StringFormatUtils.keyWordTitleToRawTitle(i.name ?? ""), // 清理HTML标签
-  realWord: i.value ?? ""
-));
-  }
-  return list;
-}
 
   ///搜索请求
   ///keyword 搜索的词
